ThreadFix 1.0 Beta 22
- Scan checks now take into account scans on the queue when checking dates
- Remote Provider scan imports and Defect Tracker updates can now be scheduled to run every day
- Snort rules have been improved with help from Sourcefire
- Qualys support has been added for the EU region
- Burp vulnerability names have been updated to match the Burp XML names
- ThreadFix now runs on Tomcat 7 and Servlet 3.0

+++ src/main/java/com/denimgroup/threadfix/data/dao/hibernate/HibernateScanDao.java
@@ -30,6 +30,7 @@ import org.hibernate.criterion.Projections;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.denimgroup.threadfix.data.dao.ScanDao;
 import com.denimgroup.threadfix.data.entities.Finding;
@@ -75,6 +76,7 @@ public class HibernateScanDao implements ScanDao {
 	}
 
 	@Override
+	@Transactional
 	public void saveOrUpdate(Scan scan) {
 		sessionFactory.getCurrentSession().saveOrUpdate(scan);
 	}
@@ -213,12 +215,15 @@ public class HibernateScanDao implements ScanDao {
 					  .setInteger("scan", scan.getId())
 					  .executeUpdate();
 		
-		sessionFactory.getCurrentSession()
-					  .createQuery("delete from SurfaceLocation " +
-					  		"where id in (:ids)")
-					  .setParameterList("ids", surfaceLocationIds)
-					  .executeUpdate();
+		if (surfaceLocationIds != null && surfaceLocationIds.size() > 0) {
+			sessionFactory.getCurrentSession()
+						  .createQuery("delete from SurfaceLocation " +
+						  		"where id in (:ids)")
+						  .setParameterList("ids", surfaceLocationIds)
+						  .executeUpdate();
+		}
 		
 		sessionFactory.getCurrentSession().delete(scan);
 	}
+	
 }
